+++ lib/CHANGELOG.md
@@ -21,6 +21,8 @@
    This means that you can now, for instance, ask if an option was used, how many times it was used,
    retrieve the value from the last instance, or get the set of values of all instances, or get an
    iterator over non-options.
+ * Added an `add_shorts_from_str` method to `OptionSetEx`, which allows adding multiple short
+   options in one go using a string.
  * Removed the `gong_option_set` macro that constructed an `OptionSetEx`, since there was very
    little point to it, with virtually no difference to creating a raw object. It was a legacy
    hangover.